+++ packages/components-vue/src/examples/TableExampleSorting.vue
@@ -30,20 +30,20 @@ const headSorting: HeadSorting[] = [
 
 type DataSorting = {
   col1: string;
-  col2: string;
+  col2: number;
   col3: string;
 };
 
 const dataSorting: DataSorting[] = [
   {
-    col1: 'Name A',
-    col2: '9',
-    col3: '01.06.2021',
+    col1: 'Name Z',
+    col2: 1,
+    col3: '24.06.2021',
   },
   {
-    col1: 'Name Z',
-    col2: '1',
-    col3: '24.06.2021',
+    col1: 'Name A',
+    col2: 9,
+    col3: '01.06.2021',
   },
 ];
 
@@ -52,10 +52,11 @@ const data = ref(dataSorting);
 
 const onUpdate = (e: CustomEvent<TableUpdateEventDetail>): void => {
   const { id, direction } = e.detail as TableUpdateEventDetail & { id: keyof DataSorting };
-  head.value = head.value.map((item) => ({ ...item, active: false, ...(item.id === id && e) }));
-  data.value = [...data.value].sort((a, b) =>
-    direction === 'asc' ? a[id].localeCompare(b[id]) : b[id].localeCompare(a[id])
-  );
+  head.value = head.value.map((item) => ({ ...item, active: false, ...(item.id === id && e.detail) }));
+  data.value = [...data.value].sort((a, b) => {
+    const result = String(a[id]).localeCompare(String(b[id]), undefined, { numeric: true });
+    return direction === 'asc' ? result : -result;
+  });
 };
 </script>