+++ packages/storefront/src/app/(main)/components/input-week/input-week.meta.ts
@@ -1,6 +1,10 @@
 import { componentMeta } from '@porsche-design-system/component-meta';
 import { inputWeekExampleForm } from '@porsche-design-system/shared/examples';
-import Accessibility from '@/app/(main)/components/input-week/accessibility/page.mdx';
+import { hiddenLabelWithoutAccessibleNameA11yExample } from '@/app/(main)/components/input-week/accessibility/examples/hidden-label-without-accessible-name/example';
+import { placeholderUsedAsTheOnlyLabelA11yExample } from '@/app/(main)/components/input-week/accessibility/examples/placeholder-used-as-the-only-label/example';
+import { validationFeedbackViaStateAndMessageApiA11yExample } from '@/app/(main)/components/input-week/accessibility/examples/validation-feedback-via-state-and-message-api/example';
+import AccessibilityOverview from '@/app/(main)/components/input-week/accessibility/overview.mdx';
+import AccessibilityTests from '@/app/(main)/components/input-week/accessibility/tests.mdx';
 import IntroductionDescription from '@/app/(main)/components/input-week/configurator/introduction.mdx';
 import { inputWeekSlotStories, inputWeekStory } from '@/app/(main)/components/input-week/configurator/story';
 import FormDescription from '@/app/(main)/components/input-week/examples/form/example.mdx';
@@ -30,7 +34,14 @@ export const inputWeekMeta = {
     },
   },
   usage: Usage,
-  accessibility: Accessibility,
+  accessibility: {
+    overview: AccessibilityOverview,
+    examples: {
+      placeholderUsedAsTheOnlyLabel: placeholderUsedAsTheOnlyLabelA11yExample,
+      hiddenLabelWithoutAccessibleName: hiddenLabelWithoutAccessibleNameA11yExample,
+      validationFeedbackViaStateAndMessageApi: validationFeedbackViaStateAndMessageApiA11yExample,
+    },
+    tests: AccessibilityTests,
+  },
   api: componentMeta['p-input-week'],
 } satisfies ComponentDocsMeta<'p-input-week'>;
-