+++ packages/storefront/src/app/(main)/components/table/accessibility/examples/column-with-hidden-label/example.ts
@@ -1,0 +1,28 @@
+import type { AccessibilityExample } from '@/models/accessibilityMeta';
+
+export const columnWithHiddenLabelA11yExample = {
+  name: 'Column with hidden label',
+  antiPattern: {
+    kind: 'story',
+    story: {
+      generator: () => [
+        {
+          tag: 'p-table-head-cell',
+          properties: { hideLabel: true },
+        },
+      ],
+    },
+  },
+  recommended: {
+    kind: 'story',
+    story: {
+      generator: () => [
+        {
+          tag: 'p-table-head-cell',
+          properties: { hideLabel: true },
+          children: ['Model'],
+        },
+      ],
+    },
+  },
+} satisfies AccessibilityExample;