+++ packages/storefront/src/app/(main)/components/toast/toast.meta.ts
@@ -1,6 +1,7 @@
 import { componentMeta } from '@porsche-design-system/component-meta';
 import { toastExample } from '@porsche-design-system/shared/examples';
-import Accessibility from '@/app/(main)/components/toast/accessibility/page.mdx';
+import AccessibilityOverview from '@/app/(main)/components/toast/accessibility/overview.mdx';
+import AccessibilityTests from '@/app/(main)/components/toast/accessibility/tests.mdx';
 import IntroductionDescription from '@/app/(main)/components/toast/configurator/introduction.mdx';
 import { toastStory } from '@/app/(main)/components/toast/configurator/story';
 import Usage from '@/app/(main)/components/toast/usage/page.mdx';
@@ -14,6 +15,10 @@ export const toastMeta = {
   },
   examples: {},
   usage: Usage,
-  accessibility: Accessibility,
+  accessibility: {
+    overview: AccessibilityOverview,
+    examples: {},
+    tests: AccessibilityTests,
+  },
   api: componentMeta['p-toast'],
 } satisfies ComponentDocsMeta<'p-toast'>;